+++ fabric.ts
@@ -1,4 +1,12 @@
-export { getEnv, getFabricDocument, getFabricWindow, setEnv } from './src/env';
+export {
+  getDevicePixelRatio,
+  getEnv,
+  getFabricDocument,
+  getFabricWindow,
+  setEnv,
+  setEnvFactory,
+} from './src/env';
+export type * from './src/env';
 export { cache } from './src/cache';
 export { VERSION as version, iMatrix } from './src/constants';
 export { config } from './src/config';