+++ packages/browser/src/index.ts
@@ -1,1 +1,6 @@
+import { setEnvFactory } from '@fabricjs/core';
+import { getEnv } from './env';
+
+setEnvFactory(getEnv);
+
 export * from '@fabricjs/core';